style: polish loading states and cleanup unused code
